pyproject.toml
[tool.poetry]
name = "telegram-server-monitor-bot"
version = "0.1.0"
description = "A simple python based telegram bot"
authors = ["Tony Benoy <me@tonybenoy.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
python-telegram-bot = "^13.13"
psutil = "^5.9.1"
feedparser = "^6.0.10"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.961"
isort = "^5.10.1"
pre-commit = "^2.20.0"
types-requests = "^2.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
